v1.0.1.2
- Changed ID variables from int to long

+++ OP_Engine/Characters/Army.cs
@@ -22,7 +22,7 @@ namespace OP_Engine.Characters
 
         #region Methods
 
-        public virtual Squad GetSquad(int id)
+        public virtual Squad GetSquad(long id)
         {
             foreach (Squad squad in Squads)
             {
